Refactoring - Relocate custom components to seperate file
